Objects set body surface; Added plant object
The object height and width comes from the surface that is used as the
body of the object. This could be changed in the future but for now it
helps to get different types of objects into the world.
Plant is now an object that doesn't move and is just a small green
square.